fix: added PeopleTable component, refactor structure of PersonalLink component
