+++ src/components/PeopleTable.tsx
@@ -1,0 +1,78 @@
+import PersonLink from '../components/PersonLink';
+import { Person } from '../types/Person';
+import cn from 'classnames';
+import { useLocation } from 'react-router-dom';
+
+type Props = {
+  people: Person[];
+};
+
+const PeopleTable: React.FC<Props> = ({ people }) => {
+  const location = useLocation();
+
+  const checkMotherArray = people?.filter(person => {
+    return people?.some(p => person.name === p.motherName);
+  });
+
+  const checkFatherArray = people?.filter(person => {
+    return people?.some(p => person.name === p.fatherName);
+  });
+
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-narrow is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        {people.map(person => {
+          const isActive = location.pathname === `/people/${person.slug}`;
+
+          const motherSlug = checkMotherArray?.find(
+            mother => mother.name === person.motherName,
+          );
+
+          const fatherSlug = checkFatherArray?.find(
+            father => father.name === person.fatherName,
+          );
+
+          return (
+            <tr
+              data-cy="person"
+              key={person.slug}
+              className={cn('', {
+                'has-background-warning': isActive,
+              })}
+            >
+              <td>
+                <PersonLink person={person} />
+              </td>
+
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+              <td>
+                <PersonLink person={motherSlug} />
+              </td>
+              <td>
+                <PersonLink person={fatherSlug} />
+              </td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+};
+
+export default PeopleTable;
