+++ src/pages/PeoplePage.tsx
@@ -1,8 +1,9 @@
 import { useEffect, useState } from 'react';
 import { Loader } from '../components/Loader';
-import PersonLink from '../components/PersonLink';
+
 import { getPeople } from '../api';
 import { Person } from '../types/Person';
+import PeopleTable from '../components/PeopleTable';
 
 const PeoplePage = () => {
   const [people, setPeople] = useState<Person[] | null>(null);
@@ -28,13 +29,6 @@ const PeoplePage = () => {
     fetchPeople();
   }, []);
 
-  const checkMotherArray = people?.filter(person => {
-    return people?.some(p => person.name === p.motherName);
-  });
-  const checkFatherArray = people?.filter(person => {
-    return people?.some(p => person.name === p.fatherName);
-  });
-
   return (
     <div>
       <h1 className="title">People Page</h1>
@@ -56,32 +50,7 @@ const PeoplePage = () => {
           )}
 
           {!isLoading && !error && people && people?.length > 0 && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              <thead>
-                <tr>
-                  <th>Name</th>
-                  <th>Sex</th>
-                  <th>Born</th>
-                  <th>Died</th>
-                  <th>Mother</th>
-                  <th>Father</th>
-                </tr>
-              </thead>
-
-              <tbody>
-                {people.map(person => (
-                  <PersonLink
-                    person={person}
-                    key={person.slug}
-                    checkMotherArray={checkMotherArray}
-                    checkFatherArray={checkFatherArray}
-                  />
-                ))}
-              </tbody>
-            </table>
+            <PeopleTable people={people} />
           )}
         </div>
       </div>
